chore: UI components refactor for layouts, store, richText, tooltip
