+++ packages/web/src/layout/accordion/index.ts
@@ -1,0 +1,4 @@
+export * from './accordion'
+export * from './accordionPanel'
+
+export * from './styles'
