+++ packages/web/src/layout/card.ts
@@ -2,6 +2,25 @@ import { html, customElement, property, css, CSSResult } from 'lit-element'
 import { MobxLitElement } from '@adobe/lit-mobx'
 import { cssReboot, cssBase } from '@c6o/ui-theme'
 
+/**
+ * `<c6o-card>` is a Web Component for displaying content in a card UI.
+ *
+ * ```
+ * <c6o-card width="350px">
+ *   <h2>${edition.displayName}</h2>
+ *   <p>${edition.description}</p>
+ * </c6o-card>
+ * ```
+ *
+ * Attribute  | Description
+ * -----------|------------
+ * `centered` | Set to center the content in the card
+ * `gap`      | Set the left and right margin that separates each card
+ * `width`    | Set to give each card a specific width
+ *
+ * @extends MobxLitElement
+ */
+
 @customElement('c6o-card')
 export class Card extends MobxLitElement {
 
@@ -11,7 +30,7 @@ export class Card extends MobxLitElement {
     @property({ type: String })
     gap = 'var(--md-spacing)'
 
-    @property({ type: String, attribute: 'card-width' })
+    @property({ type: String })
     width = 'auto'
 
     static get styles(): (CSSResult[] | CSSResult)[] {
