+++ packages/web/src/notification.ts
@@ -1,5 +1,11 @@
 import { NotificationElement } from '@vaadin/vaadin-notification/src/vaadin-notification'
 
+/**
+ * The container element for all notifications.
+ *
+ * @extends NotificationElement
+ */
+
 export interface Notification extends HTMLElement {
 }
