+++ packages/web/src/registerStyles.ts
@@ -1,28 +1,6 @@
 import { registerStyles, css } from '@vaadin/vaadin-themable-mixin/register-styles'
 
 // Register any styles here that you want to apply to the shadow DOM of a Vaadin component
-registerStyles('vaadin-accordion-panel', css`
-    :host([opened]) [part="content"] {
-        margin-bottom: var(--md-spacing);
-    }
-
-    [part~="summary"] {
-        cursor: var(--lumo-clickable-cursor);
-    }
-
-    [part~="summary-content"] ::slotted(*) {
-        margin: var(--md-spacing) 0;
-    }
-
-    :host([theme~="condensed"][opened]) [part="content"] {
-        margin-bottom: var(--xs-spacing);
-    }
-
-    :host([theme~="condensed"]) [part~="summary-content"] ::slotted(*) {
-        margin: var(--xs-spacing) 0;
-    }
-`)
-
 registerStyles('vaadin-context-menu-item', css`
     :host([tabindex])::before {
         opacity: 0;
