+++ packages/web/src/store/state.ts
@@ -3,6 +3,34 @@ import { MobxLitElement } from '@adobe/lit-mobx'
 import { EntityStore, EntityListStore } from '@c6o/common'
 import { cssReboot, cssBase, cssGrid, cssAnimations } from '@c6o/ui-theme'
 
+/**
+ * `<c6o-store-state>` is a Web Component that displays a loading banner while the store is being initialized,
+ * and a loading overlay when the store is 'busy'.
+ *
+ * ```
+ * <c6o-store-state loading-message="Looking for Apps..." .store=${this.appsListStore}>
+ *   <c6o-marketplace-apps-search
+ *     .store=${this.appsListStore}
+ *     @badge-clicked=${this.handleBadgeClick}
+ *   ></c6o-marketplace-apps-search>
+ * </c6o-store-state>
+ * ```
+ *
+ * Attribute   | Description
+ * ------------|------------
+ * `container` | When set, the loading banner will be wrapped in a div with some margins and max width
+ * `centered`  | When set, the slotted content will be centered
+ * `loading-message` | The message that will be displayed in the loading banner
+ * `no-busy`   | When set, an overlay will NOT be displayed when the store is 'busy'
+ * `no-init`   | When set, no loading banner will be displayed
+ *
+ * Property   | Description
+ * -----------|------------
+ * `store`    | The EntityStore that needs to be initialized
+ *
+ * @extends MobxLitElement
+ */
+
 @customElement('c6o-store-state')
 export class StoreState extends MobxLitElement {
